backend/myproject/requirements.txt
Django==5.1.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
psycopg2-binary~=2.9.10
requests==2.26.0